Added master/slave mapper capabilities
mappers atrtup as slaves and check for master, if there is no master the slave
will become master. if there already is a master then the slave will request
the @nanites state from the master and update its own @nanites as well as update
its persistent store of nanites.

If the master dies, the slave will promote itself to master within 6 seconds
when the old master starts up again it will start as a slave and stay that way
you can keep killing the master, the slave will promote itself and then start
the old master again and it will do the right thing.