big refactoring for security. Now agents listen on queues with crypto-strong names and they only share the hasehs queue names with the mapper. Any other agents in the system will only ever know their user name they used to  log into rabbitmq but will not know the name of the queue to send them ops on.
Also when an agent makes a request for an op, he creates a temporary queue
with a strong hashed name for the return results to come to. This way the
worker agents still don't know the requestors true queue name. The return
address queue is only temporary and is only used once and then remopved when
all results for said op are collected.

This system gives us better security then we had in XMPP since the mapper is
the only thing that knows an agents true address  Temporary queues only take
4bytes of ram in rabbitmq and are auto deleted after they are no longer in use.